Extract terminal copy mode model
